wip: move some files into core
