+++ app/src/main/java/com/chesire/malime/util/PeriodicUpdateHelper.kt
@@ -6,6 +6,7 @@ import android.content.ComponentName
 import android.content.Context
 import java.util.concurrent.TimeUnit
 
+// TODO: This should be moved into Core
 class PeriodicUpdateHelper {
     private val schedulerId = 137
