+++ app/src/main/java/com/chesire/malime/view/login/kitsu/KitsuLoginViewModel.kt
@@ -10,7 +10,7 @@ import com.chesire.malime.R
 import com.chesire.malime.kitsu.KitsuManagerFactory
 import com.chesire.malime.kitsu.models.LoginResponse
 import com.chesire.malime.util.SharedPref
-import com.chesire.malime.util.SupportedService
+import com.chesire.malime.core.SupportedService
 import com.chesire.malime.view.login.LoginModel
 import com.chesire.malime.view.login.LoginStatus
 import io.reactivex.Scheduler
