+++ app/src/main/java/com/chesire/malime/view/login/mal/MalLoginViewModel.kt
@@ -9,7 +9,7 @@ import android.support.customtabs.CustomTabsIntent
 import com.chesire.malime.R
 import com.chesire.malime.mal.MalManagerFactory
 import com.chesire.malime.util.SharedPref
-import com.chesire.malime.util.SupportedService
+import com.chesire.malime.core.SupportedService
 import com.chesire.malime.view.login.LoginModel
 import com.chesire.malime.view.login.LoginStatus
 import io.reactivex.Scheduler
