+++ malime.core/src/main/java/com/chesire/malime/core/repositories/RepositoryManager.kt
@@ -1,0 +1,9 @@
+package com.chesire.malime.core.repositories
+
+class RepositoryManager() {
+    // getManagerFor(service: SupportedService) {
+    //   if (service == *) {
+    //     return instance
+    //   }
+    // }
+}
