+++ malime.kitsu/src/main/java/com/chesire/malime/kitsu/repositories/KitsuLibrary.kt
@@ -1,23 +1,26 @@
 package com.chesire.malime.kitsu.repositories
 
 import android.content.Context
+import com.chesire.malime.core.repositories.Repository
+import com.chesire.malime.core.models.MalimeModel
 import com.chesire.malime.kitsu.api.KitsuManager
-import com.chesire.malime.kitsu.models.KitsuItem
 import com.chesire.malime.kitsu.room.KitsuDatabase
 import io.reactivex.Observable
-import io.reactivex.schedulers.Schedulers
-import timber.log.Timber
+import io.reactivex.Single
 
 class KitsuLibrary(
     context: Context,
     private val kitsuManager: KitsuManager,
     private val kitsuDb: KitsuDatabase = KitsuDatabase.getInstance(context)
-) {
-    fun getLibrary(): Observable<List<KitsuItem>> {
-        return kitsuDb.kitsuDao().getAll().toObservable()
+): Repository {
+    override fun observeLibrary(): Observable<List<MalimeModel>> {
+        return getLibraryFromDb()
     }
 
-    fun updateLibrary() {
+    override fun updateLibraryFromApi(): Single<Boolean> {
+        // need to return something to observe, so can pickup errors
+        /*
+
         kitsuManager.getUserLibrary()
             .subscribeOn(Schedulers.io())
             .subscribe(
@@ -28,5 +31,14 @@ class KitsuLibrary(
                 {
                     Timber.e("Error")
                 })
+         */
+    }
+
+    override fun clearLocalLibrary() {
+        // dbitem.clear
+    }
+
+    private fun getLibraryFromDb(): Observable<List<MalimeModel>> {
+        //return kitsuDb.kitsuDao().getAll().toObservable()
     }
 }
